Add a new option to preserve the APN configuration file

+++ systemd/server_main.py
@@ -166,7 +166,10 @@ def resolve_boot_apn():
         return None
     with open(apn_json) as apn_creds:
         apn = json.load(apn_creds)
-    os.remove(apn_json)
+    if 'PRESERVE_APN' not in os.environ or os.environ['PRESERVE_APN'] != '1':
+        os.remove(apn_json)
+    else:
+        logger.info('Preserving the APN file[%s]' % apn_json)
     return apn
